+++ src/service/walletProvider/metaMask/index.ts
@@ -1,0 +1,3 @@
+export * from './metamask.wallet.provider'
+
+