+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [4.0.0] - 2023.09.25
+### Added
+- Leveraging Extension Ecosystem a specialised Wallet Provider type of extensions was added.
+- `tatumSdk.walletProvider.use(type: TatumSdkWalletProvider)` method was added.
+
+### Changed
+- **[BREAKING CHANGE]** `tatumSdk.walletProvider.metaMask` was replaced with `tatumSdk.walletProvider.use(MetaMask)`.
+- **[BREAKING CHANGE]** `tatumSdk.walletProvider.metaMask.connect()` was replaced with `tatumSdk.walletProvider.use(MetaMask).getWallet()`.
+- **[BREAKING CHANGE]** `tatumSdk.walletProvider.metaMask.customPayload()` was replaced with `tatumSdk.walletProvider.use(MetaMask).signAndBroadcast()`.
+
 ## [3.1.8] - 2023.09.22
 ### Added
 - Capability to initialize `TatumSDK` with various extensions from Extension Ecosystem using the init() method and `TatumConfig.configureExtensions`.